+++ src/services/user/services/AdminUsers.js
@@ -11,6 +11,7 @@ const {
 	hasSchoolPermission,
 } = require('../../../hooks');
 const { adminUsersCheckUniqueAccount } = require('../hooks/userService');
+const { blockDisposableEmail } = require('../../../hooks');
 
 const { userModel } = require('../model');
 
@@ -144,7 +145,7 @@ const adminHookGenerator = (kind) => ({
 		all: [authenticate('jwt')],
 		find: [hasSchoolPermission(`${kind}_LIST`)],
 		get: [hasSchoolPermission(`${kind}_LIST`)],
-		create: [hasSchoolPermission(`${kind}_CREATE`), adminUsersCheckUniqueAccount],
+		create: [hasSchoolPermission(`${kind}_CREATE`), adminUsersCheckUniqueAccount, blockDisposableEmail('email')],
 		update: [hasSchoolPermission(`${kind}_EDIT`)],
 		patch: [hasSchoolPermission(`${kind}_EDIT`)],
 		remove: [hasSchoolPermission(`${kind}_DELETE`)],